+++ src/editor-service/local-editor-service.ts
@@ -65,10 +65,11 @@ export class LocalEditorService extends EditorService {
     }
     if (location.kind === 'tagName' || location.kind === 'text') {
       const elements =
-          Array.from(document.getByKind('element')).filter(e => e.tagName);
+          Array.from(document.getByKind('element', {imported: true}))
+              .filter((e) => e.tagName);
       return {
         kind: 'element-tags',
-        elements: elements.map(e => {
+        elements: elements.map((e) => {
           const attributesSpace = e.attributes.length > 0 ? ' ' : '';
           return {
             tagname: e.tagName!,
@@ -80,7 +81,8 @@ export class LocalEditorService extends EditorService {
         })
       };
     } else if (location.kind === 'attribute') {
-      const elements = document.getById('element', location.element.nodeName);
+      const elements = document.getById(
+          'element', location.element.nodeName, {imported: true});
       let attributes: AttributeCompletion[] = [];
       for (const element of elements) {
         // A map from the inheritedFrom to a sort prefix. Note that
@@ -95,7 +97,7 @@ export class LocalEditorService extends EditorService {
           sortPrefixes.set(element.extends, 'ccc-');
         }
         const elementAttributes: AttributeCompletion[] =
-            element.attributes.map(p => {
+            element.attributes.map((p) => {
               const sortKey =
                   (sortPrefixes.get(p.inheritedFrom) || `ddd-`) + p.name;
               return {
@@ -127,7 +129,7 @@ export class LocalEditorService extends EditorService {
   async getWarningsForFile(localPath: string): Promise<Warning[]> {
     try {
       const doc = await this._analyzer.analyze(localPath);
-      return doc.getWarnings();
+      return doc.getWarnings({imported: false});
     } catch (e) {
       // This might happen if, e.g. `localPath` has a parse error. In that case
       // we can't construct a valid Document, but we might still be able to give
@@ -152,15 +154,17 @@ export class LocalEditorService extends EditorService {
       return;
     }
     if (location.kind === 'tagName') {
-      return document.getOnlyAtId('element', location.element.nodeName);
+      return document.getOnlyAtId(
+          'element', location.element.nodeName, {imported: true});
     } else if (location.kind === 'attribute') {
-      const elements = document.getById('element', location.element.nodeName);
+      const elements = document.getById(
+          'element', location.element.nodeName, {imported: true});
       if (elements.size === 0) {
         return;
       }
 
       return concatMap(elements, (el) => el.attributes)
-          .find(at => at.name === location.attribute);
+          .find((at) => at.name === location.attribute);
     }
   }