+++ src/model/inline-document.ts
@@ -104,7 +104,7 @@ function isLocationInfo(loc: (parse5.LocationInfo|parse5.ElementLocationInfo)):
 export function getLocationOffsetOfStartOfTextContent(node: ASTNode):
     LocationOffset {
   const childNodes = node.childNodes || [];
-  const firstChildNodeWithLocation = childNodes.find(n => !!n.__location);
+  const firstChildNodeWithLocation = childNodes.find((n) => !!n.__location);
   const bestLocation = firstChildNodeWithLocation ?
       firstChildNodeWithLocation.__location :
       node.__location;